Move runHandler to dev/common
