+++ pkg/dev/common/handler_test.go
@@ -1,4 +1,4 @@
-package kubedev
+package common
 
 import (
 	"context"
